+++ app/src/main/java/com/daffodyl/medialibrary/models/BoardGame.kt
@@ -1,10 +1,10 @@
 package com.daffodyl.medialibrary.models
 
 data class BoardGame(
-    val id: Int,
+    val id: Long,
     val title: String,
-    val minPlayers: Int,
-    val maxPlayers: Int,
+    val minPlayers: Long,
+    val maxPlayers: Long,
     val genre: String,
     val notes: String
 )
