+++ app/src/main/java/com/daffodyl/medialibrary/models/Book.kt
@@ -1,11 +1,11 @@
 package com.daffodyl.medialibrary.models
 
 data class Book(
-    val id: Int,
+    val id: Long,
     val title: String,
     val author: String,
     val format: String,
-    val numPages: Int,
+    val numPages: Long,
     val genre: String,
     val notes: String
 )
